Extract KaitaiStreamBase containing utility methods.
